Remove self from functions where it was not used
There are a handful of function that were using `self` but did not
reference `self` in the function body. This coverts them to either an
associated function or moves the function completely out of the
implementation.
